requirements.txt
python-telegram-bot
tinydb
Flask
requests
datetime
google-cloud-translate
googletrans==3.1.0a0